Initial revision of the feed_manager extension
This extension allows you to maintain "replica" data for a Resilient
system.  The updates are performed in near real-time.

The use case for this extension is for cases where you want to
run business intelligence queries from other tools.  It
hopefully eliminates the desire to query the Resilient database
directly (since the Resilient DB schema is not documented and
can change at any time).

The data can be written to any or all of the following:

- ODBC data source (e.g. PostgreSQL)
- SQLite file
- Local directory

The content is consistent with the Resilient type/field
semantic model and will include custom fields.

Newly added fields are created in the destination databases
(for ODBC/SQLite destinations).

Current limitations:
- Tests have not yet been written.
- Only SQLite and PostgreSQL SQL dialects are currently supported.