INT-825 QA Changes
INT-1080 Added a app.config value for specify a number of approved images rather than one defined one. Also exposed a function field to allow users to specify an operation to perform which is then validated to ensure it is an approved operation

INT-1078 Now returning the operation done aswell as the image used in one column on the datatable. If an operation was specifed the two are divided with the `:` character. If no operation it just shows the image.