INT-825 Timestamp and UI change
INT-1080 Added  ability to not only do 2 volatility commands, but as many as you want. A config {image}_approved_operations is now exposed as a csv list you can use for operations which you intend to run with that image.
INT-1143 Added operation done to the note, same way we do in datatable
INT-1144 Added a timestamp to the resulting attachment