INT-2064 Support for offline xmlmime and date handling
For the xmlmime:
A URL is hit to setup the DLP SOAP Client. This URL is unaccessible to anyone who doesnt have public network access. To support this scenario, added a check to see if the user already provides a mime file in the `data` dir and if they did, use that as a cached xmlmime. Otherwise hit the usual location.

For the Date Handling:
If the DLP is a network event, the event_date or detection_date won't be available. Additionally in edge cases the time of the DLP server will be different to our Resilient server time. For this case take both event_date and detection_date and provided the lower of the two values for start_date and the higher for detection_date..
If neither are provided use the current time for both values (represented as 0)