Fixing Joe Sandbox jbxapi to version 2.10.1
The Joe Sandbox integration is not compatible with the latest 3.x version of the jbxapi, so the jbxapi should get fixed to the latest 2.x version. This would also work for on-prem customers according to the jbxapi documentation on GitHib: https://github.com/joesecurity/jbxapi
I have successfully tested the integration with jbxapi version 2.10.1 in my demo environment together with the Joe Sandbox Cloud Pro SaaS solution.