Update dt_utils_get_row.py
current function doesn't allow getting a row from a menu as the function to get the row_id is missing. Changing so that if you put a 0 then it will use current row as the context which is consistent with delete row

This will allow more flexible use to pull out a field from a data table using the context menu e.g. if I want to take and put into an artifact table for instance